added scripts
